pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-api-contract-tester"
dynamic = ["version"]
description = "Local-first CLI to validate API contracts against traffic"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "jsonschema>=4.0",
  "cryptography>=42.0",
]

[project.scripts]
contract-tester = "contract_tester.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "contract_tester.__version__"}

[tool.ruff]
line-length = 100
target-version = "py39"
